Broke up PosterGallery into further components (Added PosterRow) and changed some structuring to receive props from the Content component. Also edited those routes a little bit and added a few new ones in the Content component.
