+++ client/public/sass/components/_posterrow.scss
@@ -4,17 +4,17 @@
   justify-content: space-evenly;
 }
 
-.posterGallery {
+.posterRow {
   display: block;
   margin-top: 0;
 }
 
-.posterGalleryH3 {
+.posterRowH3 {
   text-align: left;
   margin: 58px;
   margin-bottom: 20px;
 }
 
-.posterGalleryPrices {
+.posterRowPrices {
   margin-bottom: 50px;
 }
