+++ client/public/sass/style.scss
@@ -13,6 +13,6 @@
   flex: 1 0 auto;
 }
 
-@import 'components/postergallery';
+@import 'components/posterrow';
 @import 'components/posterdetail';
 @import 'components/poster';
