+++ client/src/PosterGallery.tsx
@@ -1,47 +1,16 @@
 import React, {useState, useEffect} from 'react'
-import PosterTile from './PosterTile'
+import PosterRow from './PosterRow'
 
-import { Poster } from './dec';
+export interface PosterRowProps {
+    rowCategoryOne: string;
+    rowCategoryTwo: string;
+}
 
-
-function PosterGallery() {
-    const [posterArray, setPosterArray] = useState<Array<Poster>>([])
-
-    // Populate posterArray
-    useEffect(() => {
-        let newImageArray: Array<Poster> = [];
-        for (let index = 0; index < 8; index++) {
-        newImageArray.push({imageURL: 'http://placekitten.com/200/200'});
-        }
-        console.log(newImageArray)
-        setPosterArray(newImageArray);  
-        console.log(posterArray);
-    },[])
-    
+const PosterGallery: React.FC<PosterRowProps> = (props) => {
     return(
         <div>
-            <h3 className="posterGalleryH3">Bestsellers</h3>
-            <div className="postersDiv">
-                {
-                posterArray.map((poster,i) => (
-                    <div className="posterGallery">
-                        <PosterTile key={i} imageURL={poster.imageURL} />
-                        <h3 className="posterGalleryPrices">Price</h3>
-                    </div>
-                    ))
-                }
-            </div>
-            <h3 className="posterGalleryH3">Featured</h3>
-            <div className="postersDiv">
-                {
-                posterArray.map((poster,i) => (
-                    <div className="posterGallery">
-                        <PosterTile key={i} imageURL={poster.imageURL} />
-                        <h3 className="posterGalleryPrices">Price</h3>
-                    </div>
-                    ))
-                }
-            </div>
+            <PosterRow rowCategory={props.rowCategoryOne} />
+            <PosterRow rowCategory={props.rowCategoryTwo} />
         </div>
     )
 }
