+++ client/src/css/style.css
@@ -89,18 +89,18 @@ nav a:hover {
   justify-content: space-evenly;
 }
 
-.posterGallery {
+.posterRow {
   display: block;
   margin-top: 0;
 }
 
-.posterGalleryH3 {
+.posterRowH3 {
   text-align: left;
   margin: 58px;
   margin-bottom: 20px;
 }
 
-.posterGalleryPrices {
+.posterRowPrices {
   margin-bottom: 50px;
 }
